Extract quiche-client into a lib that can be used programmatically
